requirements.txt
annotated-types==0.7.0anyio==4.8.0blinker==1.9.0certifi==2025.1.31click==8.1.8colorama==0.4.6exceptiongroup==1.2.2fastapi==0.115.11Flask==3.1.0greenlet==3.1.1h11==0.14.0httpcore==1.0.7httptools==0.6.4httpx==0.28.1idna==3.10iniconfig==2.0.0itsdangerous==2.2.0Jinja2==3.1.6MarkupSafe==3.0.2MyApplication==0.1.0packaging==24.2pluggy==1.5.0psycopg2-binary==2.9.10pydantic==2.10.6pydantic_core==2.27.2pytest==8.3.5python-dotenv==1.0.1PyYAML==6.0.2sniffio==1.3.1SQLAlchemy==2.0.39starlette==0.46.1tomli==2.2.1typing_extensions==4.12.2uvicorn==0.34.0watchfiles==1.0.4websockets==15.0.1Werkzeug==3.1.3